casts tick modulo

+++ src/main.cpp
@@ -227,7 +227,7 @@ void loop()
         stop(note);
       }
     }
-    else if ((int)tick % 12 == swing)
+    else if ((int) (tick % 12) == swing)
     {
       for (Note note : shift_grid[tickToEighthNote(tick) % 8])
       {
@@ -259,7 +259,7 @@ void loop()
         }
       }
     }
-    else if (tick % 12 == swing)
+    else if ((int) (tick % 12) == swing)
     {
       for (Note note : shift_grid[tickToEighthNote(tick) % 8])
       {
